added timestamp query to ts1
this commit adds the first iteration of the timestamp query.
This feature allows users to query out a single timestamp
from the model run. Doing so allows them to query line profile, it
also saves A LOT of computation time when user is querying from large
collection of hdf files.